requirements.txt
djangorestframework==3.14.0
django-qr-code==3.1.1
psycopg2-binary==2.9.7
Pillow==10.0.0